+++ tests/unit/_bootstrap.php
@@ -1,4 +1,0 @@
-<?php
-
-require_once __DIR__.'/../../vendor/yiisoft/yii2/Yii.php';
-require_once __DIR__.'/../../vendor/craftcms/cms/src/Craft.php';